[FIXED JENKINS-15886] Make build selector parameter not add to Env
Created a new BuildSelectorParameterValue that stores the selector from
parameter input which does not expose the the value to the Enviroment.
Changed ParameterizedBuildSelector to retreived this information from
the ParametersAction rather than the enviroment.
Testcases modified and case added for when there is no parameter name
matching the that configured in builder.

These changes will break the when using Rebuild plugin with
this parameter included in the job.