Fingerprint only if any fingerprints already exist.
In Jenkins installations that don't use fingerprints (with the express
goal to not offer e.g. fingerprint-related sidepanel items), it makes
no sense for the Copy Artifact plugin to introduce them.

This change only creates fingerprints in the FingerprintingCopyMethod if
any fingerprints already exist in the Jenkins installation, otherwise,
no fingerprints are created.