Revert "Bump maven-plugin from 3.16 to 3.21" (#181)
* Revert "Bump maven-plugin from 3.16 to 3.21"

* Explain maven plugin version pin in a comment

https://github.com/jenkinsci/bom/pull/1978 notes that the upgrade of
the optional dependency on the maven plugin from 3.16 to 3.21 caused
the Jenkins plugin bill of materials to need an explicit inclusion of
the Jenkins maven plugin into the plugin BOM.

That's not a desired change because the Jenkins maven plugin tests do
not pass when run in the Jenkins plugin BOM test suites.  We don't
want to invest the effort in the Jenkins maven plugin tests to make
them work in the Jenkins plugin BOM.

The https://github.com/jenkinsci/copyartifact-plugin/pull/174
dependabot proposed upgrade is being reverted to avoid the problem.
When dependabot next proposes an upgrade of this dependency, it can be
closed or evaluated as an incremental build in the Jenkins plugin BOM
before the upgrade is approved.