Skip CopyArtifactTest#testSymlinks tests on Windows if no symlink support (#259)
Windows 10 and Windows 11 do not allow symbolic links to be created
unless the permission is granted to the user in group policy and
the Windows developer mode is enabled.

https://gist.github.com/huenisys/1efb64e57c37cfab7054c65702588fce
provides one description.

https://learn.microsoft.com/en-us/previous-versions/windows/it-pro/windows-10/security/threat-protection/security-policy-settings/create-symbolic-links
warns that it is a security risk to allow symbolic links in applications
that are not designed to handle them.